fix: Fix getting event emitter & add provideGetAmpUpdateStartedLogs
