+++ Balancer-Bots/Stable-Pool-Depegging/src/agent.ts
@@ -1,17 +1,7 @@
-import {
-  ethers,
-  Finding,
-  getEthersProvider,
-  Initialize,
-  HandleBlock,
-  BlockEvent,
-  HandleTransaction,
-  TransactionEvent,
-} from "forta-agent";
+import { ethers, Finding, getEthersProvider, Initialize, HandleBlock, BlockEvent } from "forta-agent";
 import { NetworkManager } from "forta-agent-tools";
-import { AMP_UPDATE_STARTED_ABI } from "./constants";
 import { createAbsoluteThresholdFinding, createPercentageThresholdFinding } from "./finding";
-import { NetworkData, toBn } from "./utils";
+import { NetworkData, provideGetAmpUpdateStartedLogs, toBn } from "./utils";
 import CONFIG from "./agent.config";
 import BigNumber from "bignumber.js";
 
@@ -32,33 +22,19 @@ export const provideHandleBlock = (
   networkManager: NetworkManager<NetworkData>,
   provider: ethers.providers.Provider
 ): HandleBlock => {
-  const stablePoolIface = new ethers.utils.Interface([AMP_UPDATE_STARTED_ABI]);
-  const topics = [stablePoolIface.getEventTopic("AmpUpdateStarted")];
+  const getAmpUpdateStartedLogs = provideGetAmpUpdateStartedLogs(networkManager, provider);
 
   return async (blockEvent: BlockEvent): Promise<Finding[]> => {
-    const encodedLogs = await provider.getLogs({
-      topics,
-      fromBlock: blockEvent.blockNumber,
-      toBlock: blockEvent.blockNumber,
-    });
+    const logs = await getAmpUpdateStartedLogs(blockEvent.blockNumber);
 
-    if (!encodedLogs.length) return [];
-
-    const stablePoolAddresses = networkManager.get("stablePoolAddresses");
     const absoluteThreshold = networkManager.get("absoluteThreshold");
     const percentageThreshold = networkManager.get("percentageThreshold");
 
-    const logs = encodedLogs
-      .filter((log) => stablePoolAddresses.includes(log.address.toLowerCase()))
-      .map((log) => stablePoolIface.parseLog(log));
-
     const findings: Finding[] = [];
 
-    logs.forEach((log, idx) => {
-      const poolAddress = encodedLogs[idx].address;
-
+    logs.forEach((log) => {
       if (absoluteThreshold !== undefined && log.args.endValue.lte(absoluteThreshold)) {
-        findings.push(createAbsoluteThresholdFinding(poolAddress, log.args.endValue));
+        findings.push(createAbsoluteThresholdFinding(log.emitter, log.args.endValue));
       }
 
       if (percentageThreshold !== undefined) {
@@ -68,7 +44,7 @@ export const provideHandleBlock = (
         const decreasePercentage = endValue.minus(startValue).div(endValue).shiftedBy(2);
 
         if (decreasePercentage.gte(percentageThreshold)) {
-          findings.push(createPercentageThresholdFinding(poolAddress, log.args.endValue, decreasePercentage));
+          findings.push(createPercentageThresholdFinding(log.emitter, log.args.endValue, decreasePercentage));
         }
       }
     });
